reset builds

+++ add-version.js
@@ -8,6 +8,14 @@ const dockerfile = fs
   .replace('{{METEOR_VERSION}}', version)
 
 fs.mkdirSync(`${folder}/${version}`)
-fs.unlinkSync(`${folder}/latest/Dockerfile`)
-fs.writeFileSync(`${folder}/latest/Dockerfile`, dockerfile)
 fs.writeFileSync(`${folder}/${version}/Dockerfile`, dockerfile)
+
+if (
+  version.indexOf('rc') === -1 &&
+  version.indexOf('canary') === -1 &&
+  version.indexOf('beta') === -1 &&
+  version.indexOf('alpha') === -1) {
+  // official release, update latest
+  fs.unlinkSync(`${folder}/latest/Dockerfile`)
+  fs.writeFileSync(`${folder}/latest/Dockerfile`, dockerfile)
+}
